% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_unique_colors.R
\name{determine_unique_colors}
\alias{determine_unique_colors}
\title{determine_unique_colors}
\usage{
determine_unique_colors(image)
}
\arguments{
\item{image}{}
}
\description{
determine_unique_colors
}
