% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geopalette-package.R
\docType{package}
\name{geopalette-package}
\alias{geopalette}
\alias{geopalette-package}
\title{geopalette: make color palettes from satellite images}
\description{
geopalette makes color palettes from user specified locations using satellite images.
}
\author{
\strong{Maintainer}: Charlie Bayne \email{baynec2@gmail.com} (\href{https://orcid.org/0000-0003-1870-5298}{ORCID})

}
\keyword{internal}
