% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_image_request_url.R
\name{google_image_request_url}
\alias{google_image_request_url}
\title{google_image_request_url}
\usage{
google_image_request_url(location, zoom = "city")
}
\arguments{
\item{zoom}{}
}
\description{
google_image_request_url
}
