% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_unique_colors.R
\name{plot_unique_colors}
\alias{plot_unique_colors}
\title{plot_unique_colors}
\usage{
plot_unique_colors(unique_colors)
}
\arguments{
\item{unique_colors}{list of colors to plot}
}
\value{
ggplot2 object
}
\description{
simple function to plot colors
}
\examples{

#Plotting 8  colors from culpeper palette.
culpeper = google_image_request_url("Culpeper,VA","city") \%>\%
read_image_url() \%>\%
determine_unique_colors() \%>\%
sort_colors()

culpeper_pal = create_palette(culpeper)

plot_unique_colors(culpeper_pal(8))

}
