% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_image_url.R
\name{read_image_url}
\alias{read_image_url}
\title{read_image_url}
\usage{
read_image_url(google_image_request_url)
}
\arguments{
\item{google_image_request_url}{url to get image.}
}
\value{
png image in R
}
\description{
reads image  into R from url passed to google API
}
\examples{
# This will print the image associate with the image request url
read_image_url("https://maps.googleapis.com/maps/api/staticmap?center=San\%20Diego,CA&zoom=10&&maptype=satellite&size=400x400&key=AIzaSyA7l6m8iwPSIMlb3vFNK2TAKFSy7Is7N3c")
}
