% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_geopalette.R
\name{scale_fill_geopalette}
\alias{scale_fill_geopalette}
\title{scale_fill_geopalette}
\usage{
scale_fill_geopalette(
  location,
  zoom = "city",
  discrete = TRUE,
  reverse = FALSE,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
scale_fill_geopalette
}
