% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_image.R
\name{show_image}
\alias{show_image}
\title{show_image}
\usage{
show_image(image)
}
\arguments{
\item{image}{png image read into R. intended to be the output of read_image_url()}
}
\value{
this returns a plotted image
}
\description{
plotting png image.
}
\examples{
# This will show the image used for
google_image_request_url("Culpeper,VA","city") \%>\%
read_image_url() \%>\%
show_image()
}
