% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_unique_colors.R
\name{determine_unique_colors}
\alias{determine_unique_colors}
\title{determine_unique_colors}
\usage{
determine_unique_colors(image)
}
\arguments{
\item{image}{this is the image. is intended to be the output of read_image_url()}
}
\value{
list of unique colors
}
\description{
Determining unique colors from a png image in R.
}
\examples{

# determining the unique colors from Culpeper,VA image.
culpeper = google_image_request_url("Culpeper,VA","city") \%>\%
read_image_url() \%>\%
determine_unique_colors()
}
