% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_geopalette.R
\name{scale_color_geopalette}
\alias{scale_color_geopalette}
\title{scale_color_geopalette}
\usage{
scale_color_geopalette(
  location,
  zoom = "city",
  discrete = TRUE,
  reverse = FALSE,
  ...
)
}
\arguments{
\item{discrete}{}
}
\description{
scale_color_geopalette
}
